requirements.txt
Flask==2.1.0
requests==2.27.1
spotipy==2.19.0
matplotlib==3.5.1
python-dotenv==0.19.2
